details of the record formats
What's in the NSRL?

The NSRL RDS contains metadata on computer files which can be used to uniquely identify the files and their provenance. For each file in the NSRL collection, the following data are published:
- Cryptographic hash values (MD5 and SHA-1) of the file's content. These uniquely identify the file even if, for example, it has been renamed.
- Data about the file's origin, including the software package(s) containing the file and the manufacturer of the package.
- Other data about the file, including its original name and size.

The RDS is distributed as a collection of UTF-8 text files on a set of 4 CDs. Each file contains a set of records, one record per line, consisting of comma-separated fields. Details of the record formats can be found here (PDF file).

(https://www.nist.gov/software-quality-group/nsrl-introduction)